refactor(dashboard): share the events→board projection between web kanban and TUI
Promote the pure events→board reducer to a shared, transport-neutral home
(`ouroboros.dashboard.board`) consumed by BOTH the web Kanban and the TUI.
This kills dual-reducer drift and gives the TUI per-worker provider identity
(runtime_backend per node) for free — the surface it previously lacked.

- Move `reduce_board`/`COLUMNS` to `dashboard/board.py`; `dashboard_web/kanban.py`
  becomes a thin re-export so all web imports/tests keep working unchanged.
- Add `BOARD_EVENT_TYPES` so an incremental consumer (the TUI) can bound its fold.
- TUI folds the SAME reducer output to stamp `provider` onto ac_tree nodes;
  dashboard_v3 renders a provider tag on AC/worker rows + a Provider detail row.
  No screen rewrite — pause/resume, debug inspector, logs untouched.
