+++ src/ouroboros/tui/app.py
@@ -17,6 +17,7 @@ from typing import TYPE_CHECKING, Any, ClassVar
 from textual.app import App
 from textual.binding import Binding
 
+from ouroboros.dashboard.board import BOARD_EVENT_TYPES, reduce_board
 from ouroboros.tui.events import (
     ACUpdated,
     AgentThinkingUpdated,
@@ -370,6 +371,11 @@ class OuroborosTUI(App[None]):
             self.post_message(message)
             self._update_state_from_event(event)
 
+        # Fold provider identity from the SHARED board reducer (same projection the
+        # web Kanban renders). Additive: it only annotates provider tags and never
+        # touches pause/resume/log/debug flows above.
+        self._ingest_board_event(event)
+
         # Forward raw event to debug screen
         try:
             debug_screen = self.get_screen("debug")
@@ -385,6 +391,40 @@ class OuroborosTUI(App[None]):
                 )
         except Exception:
             pass  # Screen might not be installed yet
+
+    def _ingest_board_event(self, event: BaseEvent) -> None:
+        """Fold one event into the shared board projection for provider identity.
+
+        The TUI keeps its own hierarchical ``ac_tree`` for layout, but derives
+        per-node PROVIDER (runtime_backend) from the exact same reducer the web
+        Kanban uses — so the two surfaces can never drift on who ran what. We fold
+        a bounded tail of board-relevant events (tool spam excluded via
+        ``BOARD_EVENT_TYPES``) and read the provider off each card.
+        """
+        if event.type not in BOARD_EVENT_TYPES:
+            return
+
+        data = event.data if isinstance(event.data, dict) else {}
+        self._state.board_events.append({"event_type": event.type, "payload": dict(data)})
+
+        board = reduce_board(
+            self._state.board_events,
+            execution_id=self._execution_id or None,
+        )
+        provider_by_node: dict[str, str] = {}
+        for column in board["columns"].values():
+            for card in column:
+                provider = card.get("provider")
+                node_id = card.get("id")
+                if isinstance(provider, str) and provider and isinstance(node_id, str):
+                    provider_by_node[node_id] = provider
+
+        self._state.provider_by_node = provider_by_node
+        self._state.board_providers = [p for p in board["providers"] if isinstance(p, str)]
+        # Stamp providers onto any tree nodes that already exist. Nodes created
+        # later (by queued Textual messages) get stamped at their own
+        # ``_notify_ac_tree_updated`` via ``_apply_provider_tags``.
+        self._notify_ac_tree_updated()
 
     async def _subscribe_to_events(self, context: _EventSubscriptionContext) -> None:
         """Subscribe to EventStore for live updates.
@@ -533,6 +573,9 @@ class OuroborosTUI(App[None]):
         self._state.active_tools.clear()
         self._state.tool_history.clear()
         self._state.thinking.clear()
+        self._state.board_events.clear()
+        self._state.provider_by_node.clear()
+        self._state.board_providers.clear()
         self._state.add_log("info", "tui.main", f"Monitoring execution: {execution_id}")
         # Forward to logs screen
         try:
@@ -991,8 +1034,28 @@ class OuroborosTUI(App[None]):
         if dashboard is not None and hasattr(dashboard, method_name):
             getattr(dashboard, method_name)(message)
 
+    def _apply_provider_tags(self) -> None:
+        """Stamp per-node provider onto ac_tree nodes from the shared projection.
+
+        Single choke point: whatever built/mutated the tree, the provider tag is
+        applied here right before the widget renders, keyed by node_id — so a
+        provider learned before OR after its node was created still lands.
+        """
+        nodes = self._state.ac_tree.get("nodes")
+        if not isinstance(nodes, dict) or not self._state.provider_by_node:
+            return
+        for node_id, node in nodes.items():
+            if not isinstance(node, dict):
+                continue
+            provider = self._state.provider_by_node.get(
+                node_id
+            ) or self._state.provider_by_node.get(node.get("node_id", ""))
+            if provider:
+                node["provider"] = provider
+
     def _notify_ac_tree_updated(self) -> None:
         """Notify dashboard that AC tree has been updated."""
+        self._apply_provider_tags()
         dashboard = self._get_dashboard_screen()
         if dashboard is not None and hasattr(dashboard, "_tree") and dashboard._tree is not None:
             dashboard._tree.update_tree(self._state.ac_tree)
